refactor: extract resolve module from lib.rs
Move source resolution functions and related types to src/resolve.rs:
- canonicalize_path
- ResolvedSource, ResolvedSources structs/enums
- resolve_source (main entry point)
- resolve_github_url, resolve_local_path, resolve_two_part, resolve_three_part
- prompt_save_source, get_cached_repo_commit
- list_overlays_from_cached_repo, list_overlays_from_path
- visible_subdirs, select_overlays_interactive
- resolve_from_sources_with_suggestions, fuzzy_suggest, format_not_found_error
- try_upgrade_github_source, find_matching_source, get_fuzzy_suggestions_multi_source

Also move associated test modules to resolve.rs.
