+++ src/lib.rs
@@ -17,6 +17,8 @@ mod overlay_name;
 mod overlay_repo;
 mod reference;
 mod remove;
+mod resolve;
+mod resolve;
 mod selection;
 mod sources;
 mod state;
@@ -36,6 +38,10 @@ pub(crate) use git_exclude::{
     remove_overlay_section, repair_git_exclude, update_git_exclude,
 };
 pub(crate) use remove::{remove_overlay, remove_single_overlay};
+pub(crate) use resolve::{
+    ResolvedSource, ResolvedSources, canonicalize_path, get_cached_repo_commit,
+    list_overlays_from_cached_repo, resolve_source, try_upgrade_github_source,
+};
 pub(crate) use status::{show_status, show_status_json, status_has_overlays};
 
 /// Run the CLI application.
@@ -229,910 +235,6 @@ fn show_file_diff(existing_path: &Path, overlay_path: &Path, display_path: &Path
         }
     }
     eprintln!();
-}
-
-/// Canonicalize a path and return an error with a descriptive message if it fails.
-pub(crate) fn canonicalize_path(path: &Path, description: &str) -> Result<PathBuf> {
-    path.canonicalize()
-        .with_context(|| format!("{} not found: {}", description, path.display()))
-}
-
-/// Resolved source information for applying an overlay.
-pub(crate) struct ResolvedSource {
-    /// Local path to the overlay files
-    pub path: PathBuf,
-    /// Source information for state tracking
-    pub source_info: OverlaySource,
-}
-
-/// Result of source resolution - may be a single overlay or multiple from browse mode.
-pub(crate) enum ResolvedSources {
-    /// A single resolved overlay source.
-    Single(ResolvedSource),
-    /// Multiple resolved overlay sources (from browse mode multi-select).
-    Multiple(Vec<ResolvedSource>),
-}
-
-/// Resolve a source string to a local path.
-///
-/// Resolution order:
-/// 1. GitHub URL (`https://github.com/...`) - downloads to cache, returns cached path
-/// 2. Explicit local path (`./path`, `/path`, `~/path`) - returns path directly
-/// 3. Existing local path (backward compat, warns about future `./` requirement)
-/// 4. Three-part reference (`org/repo/name`) - resolves from configured sources
-/// 5. Two-part reference (`org/repo`) - browse mode (Phase B)
-/// 6. One-part reference (`username`) - expands to `username/repo-overlays` (Phase C)
-///
-/// Find a configured source whose URL matches the given GitHub owner/repo.
-///
-/// Loads the user config and checks each source URL against the provided
-/// owner and repo. Returns the first matching source, if any.
-fn find_matching_source(owner: &str, repo: &str) -> Option<config::Source> {
-    let config = config::load_config(None).ok()?;
-    config.sources.into_iter().find(|source| {
-        source
-            .url
-            .as_deref()
-            .and_then(github::parse_remote_url)
-            .is_some_and(|(src_owner, src_repo)| src_owner == owner && src_repo == repo)
-    })
-}
-
-/// Upgrade a GitHub overlay state to `OverlayRepo` if it matches a configured source.
-///
-/// When an overlay was originally applied via a GitHub URL that points to a configured
-/// overlay repo source, this converts the state from read-only `GitHub` to editable
-/// `OverlayRepo`. The upgrade is persisted to disk.
-///
-/// Returns `true` if the state was upgraded and saved.
-///
-/// # Errors
-///
-/// Returns an error if saving the upgraded state fails.
-pub(crate) fn try_upgrade_github_source(target: &Path, state: &mut OverlayState) -> Result<bool> {
-    let (owner, gh_repo, subpath, commit) = match &state.source {
-        OverlaySource::GitHub {
-            owner,
-            repo,
-            subpath: Some(subpath),
-            commit,
-            ..
-        } => (owner.clone(), repo.clone(), subpath.clone(), commit.clone()),
-        _ => return Ok(false),
-    };
-
-    // Parse subpath as org/repo/name (3 non-empty parts)
-    let parts: Vec<&str> = subpath.split('/').collect();
-    if parts.len() != 3 || parts.iter().any(|p| p.is_empty()) {
-        return Ok(false);
-    }
-    let (org, target_repo, name) = (
-        parts[0].to_string(),
-        parts[1].to_string(),
-        parts[2].to_string(),
-    );
-
-    let Some(source) = find_matching_source(&owner, &gh_repo) else {
-        return Ok(false);
-    };
-
-    state.source = OverlaySource::overlay_repo_full(
-        org,
-        target_repo,
-        name.clone(),
-        commit,
-        ResolvedVia::Direct,
-        source.name.clone(),
-    );
-    save_overlay_state(target, state)?;
-
-    println!(
-        "{} overlay '{}' from GitHub to overlay repo source '{}' (now editable)",
-        "Upgraded".green().bold(),
-        name,
-        source.name,
-    );
-
-    Ok(true)
-}
-
-/// # Errors
-///
-/// Returns an error if:
-/// - The source doesn't match any valid format
-/// - A local path doesn't exist
-/// - GitHub fetch fails
-/// - Overlay repo is not configured (for structured references)
-/// - `source_filter` specifies an unknown source
-pub(crate) fn resolve_source(
-    source_str: &str,
-    ref_override: Option<&str>,
-    update: bool,
-    target_path: Option<&Path>,
-    source_filter: Option<&str>,
-) -> Result<ResolvedSources> {
-    debug!(
-        "resolve_source: {source_str} (ref_override={ref_override:?}, update={update}, source_filter={source_filter:?})"
-    );
-
-    // Check library first for bare names or explicit @library source filter
-    let is_bare_name = !source_str.contains('/') && !source_str.starts_with('.');
-    let is_library_filter = source_filter == Some(library::LIBRARY_SOURCE_NAME);
-    if (is_bare_name || is_library_filter)
-        && let Some(target) = target_path
-        && let Ok(library_path) = library::get_library_path(target)
-        && let overlay_path = library_path.join(source_str)
-        && overlay_path.is_dir()
-    {
-        // Canonicalize so symlinks are created with absolute paths
-        let overlay_path = overlay_path.canonicalize().unwrap_or(overlay_path);
-        debug!(
-            "resolved '{source_str}' from library at {}",
-            overlay_path.display()
-        );
-        return Ok(ResolvedSources::Single(ResolvedSource {
-            path: overlay_path,
-            source_info: state::OverlaySource::library(source_str.to_string()),
-        }));
-    }
-
-    // If --from @library was explicitly requested but overlay wasn't found, error immediately
-    if is_library_filter {
-        let library_overlays = target_path
-            .and_then(|t| library::get_library_path(t).ok())
-            .and_then(|lp| library::list_library_overlays(&lp).ok())
-            .unwrap_or_default();
-
-        if library_overlays.is_empty() {
-            bail!(
-                "No overlays found in library. Import one with: repoverlay library import <source>"
-            );
-        }
-        let names: Vec<&str> = library_overlays.iter().map(|o| o.name.as_str()).collect();
-        bail!(
-            "Overlay '{}' not found in library. Available: {}",
-            source_str,
-            names.join(", ")
-        );
-    }
-
-    // Parse input into structured reference
-    let reference = SourceReference::parse(source_str);
-    debug!("parsed reference: {reference:?}");
-
-    match reference {
-        SourceReference::GitHubUrl(url) => {
-            // Check if this GitHub URL matches a configured source — if so, resolve
-            // as an overlay repo (editable) instead of a read-only GitHub source.
-            if let Ok(github_source) = GitHubSource::parse(&url)
-                && let Some(matched) =
-                    find_matching_source(&github_source.owner, &github_source.repo)
-            {
-                // Check if the URL includes a subpath that looks like org/repo/name
-                if let Some(ref subpath) = github_source.subpath {
-                    let subpath_str = subpath.to_string_lossy().to_string();
-                    let parts: Vec<&str> =
-                        subpath_str.split('/').filter(|s| !s.is_empty()).collect();
-                    if parts.len() >= 3 {
-                        let (org, repo_name, name) = (parts[0], parts[1], parts[2]);
-                        println!(
-                            "{} Detected configured source '{}', resolving as overlay repo (editable)",
-                            "Info".cyan().bold(),
-                            matched.name,
-                        );
-                        return resolve_three_part(
-                            org,
-                            repo_name,
-                            name,
-                            target_path,
-                            source_filter,
-                            update,
-                        )
-                        .map(ResolvedSources::Single);
-                    }
-                }
-
-                // Bare repo URL (no qualifying subpath) — browse mode, but editable
-                println!(
-                    "{} Detected configured source '{}', resolving as overlay repo (editable)",
-                    "Info".cyan().bold(),
-                    matched.name,
-                );
-                return resolve_two_part(
-                    &github_source.owner,
-                    &github_source.repo,
-                    ref_override,
-                    update,
-                    target_path,
-                    source_filter,
-                    Some(&matched),
-                );
-            }
-
-            resolve_github_url(&url, ref_override, update).map(ResolvedSources::Single)
-        }
-
-        SourceReference::LocalPath {
-            path,
-            needs_prefix_warning,
-        } => {
-            resolve_local_path(&path, source_str, needs_prefix_warning).map(ResolvedSources::Single)
-        }
-
-        SourceReference::ThreePart {
-            owner,
-            repo,
-            overlay,
-        } => resolve_three_part(&owner, &repo, &overlay, target_path, source_filter, update)
-            .map(ResolvedSources::Single),
-
-        SourceReference::TwoPart { owner, repo } => resolve_two_part(
-            &owner,
-            &repo,
-            ref_override,
-            update,
-            target_path,
-            source_filter,
-            None,
-        ),
-
-        SourceReference::OnePart { username } => {
-            // Before treating as a GitHub username, check if this is an applied
-            // overlay name that could be resolved from its source. This catches
-            // cases like `switch ff-oce` where the user means an overlay name,
-            // not a GitHub user.
-            if let Some(target) = target_path {
-                let applied = state::list_applied_overlays(target).unwrap_or_default();
-                if applied.iter().any(|n| n.as_str() == username) {
-                    // Already applied — resolve from its existing source
-                    let overlay_state = state::load_overlay_state(target, &username)?;
-                    let path = match &overlay_state.source {
-                        state::OverlaySource::Library { name } => {
-                            // Library sources need repo context to resolve
-                            let lib_path = library::get_library_path(target)?;
-                            lib_path.join(name).canonicalize()?
-                        }
-                        source => {
-                            use state::SourceResolver;
-                            source.resolve_local_path()?
-                        }
-                    };
-                    return Ok(ResolvedSources::Single(ResolvedSource {
-                        path,
-                        source_info: overlay_state.source,
-                    }));
-                }
-            }
-
-            // Phase C: Expand username to username/{default_repo}
-            let default_repo = config::default_overlay_repo_name();
-            debug!("expanding one-part reference: {username} -> {username}/{default_repo}",);
-            resolve_two_part(
-                &username,
-                &default_repo,
-                ref_override,
-                update,
-                target_path,
-                source_filter,
-                None,
-            )
-        }
-    }
-}
-
-/// Resolve a GitHub URL source.
-fn resolve_github_url(
-    url: &str,
-    ref_override: Option<&str>,
-    update: bool,
-) -> Result<ResolvedSource> {
-    debug!("resolving GitHub URL: {url}");
-    let mut github_source = GitHubSource::parse(url)?;
-
-    // Apply ref override if provided
-    if let Some(ref_str) = ref_override {
-        github_source = github_source.with_ref_override(Some(ref_str))?;
-    }
-
-    // Ensure cached and get path
-    let cache = CacheManager::new()?;
-
-    let label = if update {
-        "Fetching"
-    } else {
-        "Fetching (cached)"
-    };
-    println!(
-        "{} repository: {}/{}",
-        label.blue().bold(),
-        github_source.owner,
-        github_source.repo
-    );
-
-    let cached = cache.ensure_cached(&github_source, update)?;
-
-    Ok(ResolvedSource {
-        path: cached.path,
-        source_info: OverlaySource::github(
-            url.to_string(),
-            github_source.owner,
-            github_source.repo,
-            github_source.git_ref.as_str().to_string(),
-            cached.commit,
-            github_source
-                .subpath
-                .map(|p| p.to_string_lossy().to_string()),
-        ),
-    })
-}
-
-/// Resolve a local filesystem path.
-fn resolve_local_path(
-    path: &Path,
-    original_input: &str,
-    needs_prefix_warning: bool,
-) -> Result<ResolvedSource> {
-    debug!("resolving local path: {}", path.display());
-
-    // Require `./` prefix for local paths to avoid ambiguity with overlay repo references.
-    // Absolute paths (`/...`) and home-relative paths (`~/...`) are unambiguous and allowed.
-    if needs_prefix_warning {
-        bail!(
-            "Ambiguous path '{}': use './{original_input}' to specify a local path explicitly.",
-            path.display()
-        );
-    }
-
-    if !path.exists() {
-        bail!("Overlay source not found: {}", path.display());
-    }
-
-    let canonical = path
-        .canonicalize()
-        .with_context(|| format!("Overlay source not found: {}", path.display()))?;
-
-    Ok(ResolvedSource {
-        path: canonical.clone(),
-        source_info: OverlaySource::local(canonical),
-    })
-}
-
-/// Resolve a two-part overlay reference by browsing a GitHub overlay repository.
-///
-/// This is "GitHub browse mode" - the user specified a GitHub repo containing overlays
-/// (e.g., `tylerbutler/repo-overlays`). We clone it to the GitHub cache and let them
-/// interactively select one or more overlays.
-///
-/// The overlay repo structure is: `<target_org>/<target_repo>/<overlay_name>/`
-/// For example: `microsoft/FluidFramework/vscode-setup/`
-///
-/// Note: This differs from `resolve_three_part` which uses configured overlay sources.
-/// Here we resolve directly from the cached GitHub repo, not through `SourceManager`.
-fn resolve_two_part(
-    owner: &str,
-    repo: &str,
-    ref_override: Option<&str>,
-    update: bool,
-    target_path: Option<&Path>,
-    _source_filter: Option<&str>,
-    matched_source: Option<&config::Source>,
-) -> Result<ResolvedSources> {
-    let mode = if matched_source.is_some() {
-        "overlay repo browse"
-    } else {
-        "GitHub browse"
-    };
-    debug!("resolving two-part reference ({mode} mode): {owner}/{repo}");
-
-    // Create a GitHubSource to fetch/cache the repo
-    let github_url = format!("https://github.com/{owner}/{repo}");
-    let mut github_source = GitHubSource::parse(&github_url)?;
-
-    if let Some(ref_str) = ref_override {
-        github_source = github_source.with_ref_override(Some(ref_str))?;
-    }
-
-    // Fetch/cache the repository
-    let cache = CacheManager::new()?;
-    let label = if update {
-        "Fetching"
-    } else {
-        "Fetching (cached)"
-    };
-    println!("{} repository: {}/{}", label.blue().bold(), owner, repo);
-    let cached = cache.ensure_cached(&github_source, update)?;
-
-    // List available overlays
-    let available_overlays = list_overlays_from_cached_repo(owner, repo)?;
-
-    if available_overlays.is_empty() {
-        bail!(
-            "No overlays found in {owner}/{repo}.\n\n\
-             Make sure the repository contains overlay directories in the format:\n\
-             <target-org>/<target-repo>/<overlay-name>/"
-        );
-    }
-
-    // Select overlays based on interactivity
-    let applied = target_path
-        .map(|p| list_applied_overlays(p).unwrap_or_default())
-        .unwrap_or_default();
-    let repo_identity = target_path.and_then(|p| upstream::detect_repo_identity(p).ok().flatten());
-    let selected_overlays = if is_interactive() {
-        select_overlays_interactive(
-            owner,
-            repo,
-            &available_overlays,
-            &applied,
-            repo_identity.as_ref(),
-        )?
-    } else {
-        // Non-interactive mode - error with available overlays
-        let overlay_list = available_overlays
-            .iter()
-            .map(|o| format!("  {}", o.display_bold()))
-            .collect::<Vec<_>>()
-            .join("\n");
-        bail!(
-            "No overlay specified. Available overlays in {owner}/{repo}:\n{overlay_list}\n\n\
-             Use: repoverlay apply {owner}/{repo}/<target-org>/<target-repo>/<overlay-name>"
-        );
-    };
-
-    if selected_overlays.len() == 1 {
-        println!(
-            "{} overlay: {}",
-            "Selected".green().bold(),
-            selected_overlays[0].display_bold()
-        );
-    } else {
-        println!(
-            "{} {} overlays:",
-            "Selected".green().bold(),
-            selected_overlays.len()
-        );
-        for selected in &selected_overlays {
-            println!("  - {}", selected.display_bold());
-        }
-    }
-
-    // Skip save prompt when source is already configured
-    if matched_source.is_none() {
-        prompt_save_source(owner, repo)?;
-    }
-
-    // Resolve each selected overlay to a ResolvedSource
-    let commit = get_cached_repo_commit(&cached.path).unwrap_or_else(|| "unknown".to_string());
-
-    let mut resolved_sources = Vec::with_capacity(selected_overlays.len());
-
-    for selected in &selected_overlays {
-        let overlay_path = cached
-            .path
-            .join(&selected.org)
-            .join(&selected.repo)
-            .join(&selected.name);
-
-        if !overlay_path.exists() {
-            bail!("Overlay directory not found: {}", overlay_path.display());
-        }
-
-        let source_info = if let Some(source) = matched_source {
-            // Resolve as overlay repo — editable and syncable
-            OverlaySource::overlay_repo_full(
-                selected.org.clone(),
-                selected.repo.clone(),
-                selected.name.clone(),
-                commit.clone(),
-                ResolvedVia::Direct,
-                source.name.clone(),
-            )
-        } else {
-            // Resolve as read-only GitHub source
-            let git_ref_str = github_source.git_ref.as_str().to_string();
-            OverlaySource::github(
-                github_url.clone(),
-                owner.to_string(),
-                repo.to_string(),
-                git_ref_str,
-                commit.clone(),
-                Some(selected.to_string()),
-            )
-        };
-
-        resolved_sources.push(ResolvedSource {
-            path: overlay_path,
-            source_info,
-        });
-    }
-
-    if resolved_sources.len() == 1 {
-        Ok(ResolvedSources::Single(
-            resolved_sources.into_iter().next().unwrap(),
-        ))
-    } else {
-        Ok(ResolvedSources::Multiple(resolved_sources))
-    }
-}
-
-/// Prompt the user to save a source for future use, if not already configured.
-///
-/// Skips silently if the source is already configured or the session is non-interactive.
-fn prompt_save_source(owner: &str, repo: &str) -> Result<()> {
-    if !is_interactive() {
-        return Ok(());
-    }
-
-    let mut config = config::load_config(None)?;
-    let url = format!("https://github.com/{owner}/{repo}");
-    let source_name = repo.to_string();
-
-    if config
-        .sources
-        .iter()
-        .any(|s| s.url.as_deref() == Some(&url) || s.name == source_name)
-    {
-        return Ok(());
-    }
-
-    let prompt = format!("Save {owner}/{repo} as a source for future use?");
-    let confirmed = dialoguer::Confirm::new()
-        .with_prompt(&prompt)
-        .default(true)
-        .interact()?;
-
-    if !confirmed {
-        return Ok(());
-    }
-
-    config.sources.push(config::Source {
-        name: source_name.clone(),
-        url: Some(url.clone()),
-        path: None,
-    });
-    config::save_config(&config)?;
-
-    println!(
-        "{} source '{}' ({})",
-        "Saved".green().bold(),
-        source_name,
-        url
-    );
-
-    Ok(())
-}
-
-/// Get the current commit hash from a cached repository.
-pub(crate) fn get_cached_repo_commit(repo_path: &Path) -> Option<String> {
-    use std::process::Command;
-
-    let output = Command::new("git")
-        .args(["rev-parse", "HEAD"])
-        .current_dir(repo_path)
-        .output()
-        .ok()?;
-
-    if output.status.success() {
-        Some(String::from_utf8_lossy(&output.stdout).trim().to_string())
-    } else {
-        None
-    }
-}
-
-/// List overlays from a cached GitHub repository.
-///
-/// Used by GitHub browse mode (`resolve_two_part`) to list overlays from a
-/// directly-specified GitHub repo (e.g., `tylerbutler/repo-overlays`).
-///
-/// Overlay repos have a nested structure: `target_org/target_repo/overlay_name/`
-/// Returns `AvailableOverlay` structs with org, repo, and overlay name.
-///
-/// Note: This is the same structure used by `OverlayRepoManager::list_overlays()`,
-/// but operates on the GitHub cache instead of managed overlay repositories.
-pub(crate) fn list_overlays_from_cached_repo(
-    owner: &str,
-    repo: &str,
-) -> Result<Vec<AvailableOverlay>> {
-    debug!("listing overlays from cached GitHub repo: {owner}/{repo}");
-
-    let cache = CacheManager::new()?;
-    let cache_dir = cache.cache_dir();
-    let repo_path = cache_dir.join("github").join(owner).join(repo);
-
-    if !repo_path.exists() {
-        bail!("Repository not cached: {owner}/{repo}");
-    }
-
-    list_overlays_from_path(&repo_path)
-}
-
-/// List overlays from a directory with nested org/repo/overlay structure.
-///
-/// This is the core logic for listing overlays, extracted for testability.
-fn list_overlays_from_path(repo_path: &Path) -> Result<Vec<AvailableOverlay>> {
-    let mut overlays = Vec::new();
-
-    for (org_path, org_name) in visible_subdirs(repo_path)? {
-        for (repo_dir, repo_name) in visible_subdirs(&org_path)? {
-            for (overlay_path, overlay_name) in visible_subdirs(&repo_dir)? {
-                let has_config = overlay_path.join("repoverlay.ccl").exists();
-                overlays.push(AvailableOverlay {
-                    org: org_name.clone(),
-                    repo: repo_name.clone(),
-                    name: overlay_name,
-                    has_config,
-                    flat: false,
-                });
-            }
-        }
-    }
-
-    overlays.sort_by(|a, b| (&a.org, &a.repo, &a.name).cmp(&(&b.org, &b.repo, &b.name)));
-    debug!("found {} overlays in path", overlays.len());
-    Ok(overlays)
-}
-
-/// Returns visible (non-hidden) subdirectories with their names.
-fn visible_subdirs(path: &Path) -> Result<Vec<(PathBuf, String)>> {
-    let mut results = Vec::new();
-    for entry in fs::read_dir(path)? {
-        let entry = entry?;
-        let entry_path = entry.path();
-        if !entry_path.is_dir() {
-            continue;
-        }
-        if let Some(name) = entry.file_name().to_str().filter(|n| !n.starts_with('.')) {
-            results.push((entry_path, name.to_string()));
-        }
-    }
-    Ok(results)
-}
-
-/// Present an interactive multi-select picker for overlays.
-///
-/// Already-applied overlays appear dimmed and cannot be selected.
-/// When `repo_identity` is provided, overlays targeting a different repository
-/// are shown dimmed with a "different repo" description but remain selectable.
-/// Matching overlays are listed first.
-fn select_overlays_interactive(
-    owner: &str,
-    repo: &str,
-    overlays: &[AvailableOverlay],
-    applied_overlays: &[OverlayName],
-    repo_identity: Option<&upstream::RepoIdentity>,
-) -> Result<Vec<AvailableOverlay>> {
-    use crate::selection::{FlatSelectionConfig, SelectableItem, select_flat};
-
-    // Order overlays: matching repo first, then non-matching
-    let ordered: Vec<&AvailableOverlay> = repo_identity.map_or_else(
-        || overlays.iter().collect(),
-        |identity| {
-            let (matching, non_matching): (Vec<_>, Vec<_>) = overlays
-                .iter()
-                .partition(|o| identity.matches(&o.org, &o.repo));
-            matching.into_iter().chain(non_matching).collect()
-        },
-    );
-
-    let items: Vec<SelectableItem> = ordered
-        .iter()
-        .map(|o| {
-            let already_applied = normalize_overlay_name(&o.name)
-                .ok()
-                .is_some_and(|normalized| {
-                    applied_overlays.iter().any(|n| n == normalized.as_str())
-                });
-            let is_different_repo =
-                repo_identity.is_some_and(|identity| !identity.matches(&o.org, &o.repo));
-
-            let description = if already_applied {
-                Some("already applied".into())
-            } else if is_different_repo {
-                Some("different repo".into())
-            } else {
-                None
-            };
-
-            SelectableItem {
-                id: o.to_string(),
-                label: o.to_string(),
-                description,
-                preselected: false,
-                disabled: already_applied,
-            }
-        })
-        .collect();
-
-    let result = select_flat(
-        &items,
-        &FlatSelectionConfig {
-            prompt: format!("Select overlay(s) from {owner}/{repo}:"),
-        },
-    )?;
-
-    if result.cancelled || result.selected_ids.is_empty() {
-        bail!("No overlays selected");
-    }
-
-    // Map selected IDs back to AvailableOverlay values
-    let selected: Vec<AvailableOverlay> = result
-        .selected_ids
-        .iter()
-        .filter_map(|id| overlays.iter().find(|o| o.to_string() == *id).cloned())
-        .collect();
-
-    Ok(selected)
-}
-
-/// Resolve a three-part overlay reference (`org/repo/overlay`).
-///
-/// Provides fuzzy suggestions when the overlay is not found.
-fn resolve_three_part(
-    org: &str,
-    repo: &str,
-    name: &str,
-    target_path: Option<&Path>,
-    source_filter: Option<&str>,
-    update: bool,
-) -> Result<ResolvedSource> {
-    debug!("resolving three-part reference: {org}/{repo}/{name}");
-
-    // Load config
-    let config = config::load_config(None)?;
-
-    // Detect upstream for fallback resolution
-    let upstream = target_path.and_then(|p| detect_upstream(p).ok()).flatten();
-
-    // Try multi-source resolution first if sources are configured
-    if !config.sources.is_empty() {
-        debug!(
-            "using multi-source resolution with {} sources",
-            config.sources.len()
-        );
-        return resolve_from_sources_with_suggestions(
-            &config.sources,
-            org,
-            repo,
-            name,
-            upstream.as_ref(),
-            source_filter,
-            update,
-            target_path,
-        );
-    }
-
-    // No sources configured
-    bail!(
-        "Overlay repository not configured.\n\n\
-         To apply overlays from a shared repository, first run:\n\
-         repoverlay source add <url>\n\n\
-         Or use a local path or GitHub URL instead."
-    )
-}
-
-/// Resolve an overlay from configured sources with fuzzy suggestions on failure.
-#[allow(clippy::too_many_arguments)]
-fn resolve_from_sources_with_suggestions(
-    sources: &[config::Source],
-    org: &str,
-    repo: &str,
-    name: &str,
-    upstream: Option<&upstream::UpstreamInfo>,
-    source_filter: Option<&str>,
-    update: bool,
-    repo_root: Option<&Path>,
-) -> Result<ResolvedSource> {
-    let manager = sources::SourceManager::new(sources.to_vec(), repo_root)?;
-
-    // Ensure all sources are cloned and up-to-date
-    manager.ensure_all_cloned()?;
-
-    if update {
-        println!("{} overlay sources...", "Updating".blue().bold());
-        manager.pull_all()?;
-    } else {
-        debug!("skipping overlay source update (--no-update)");
-    }
-
-    // Resolve overlay from sources
-    if let Some(resolved) = manager.resolve(org, repo, name, upstream, source_filter)? {
-        // Determine actual org/repo for state tracking
-        let via_upstream = resolved.resolved_via == state::ResolvedVia::Upstream;
-        let (actual_org, actual_repo) = match (upstream, via_upstream) {
-            (Some(up), true) => (up.org.clone(), up.repo.clone()),
-            _ => (org.to_string(), repo.to_string()),
-        };
-
-        let via_suffix = if via_upstream {
-            " (via upstream)".dimmed().to_string()
-        } else {
-            String::new()
-        };
-        let source_suffix = format!(" [{}]", resolved.source.name).cyan().to_string();
-
-        println!(
-            "{} overlay: {}/{}/{}{}{}",
-            "Resolving".blue().bold(),
-            actual_org,
-            actual_repo,
-            name,
-            via_suffix,
-            source_suffix,
-        );
-
-        return Ok(ResolvedSource {
-            path: resolved.path,
-            source_info: OverlaySource::overlay_repo_full(
-                actual_org,
-                actual_repo,
-                name.to_string(),
-                resolved.commit,
-                resolved.resolved_via,
-                resolved.source.name,
-            ),
-        });
-    }
-
-    // Overlay not found - provide fuzzy suggestions
-    let suggestions = get_fuzzy_suggestions_multi_source(&manager, org, repo, name);
-    let source_list = manager.source_names().join(", ");
-    let error_msg = format_not_found_error(org, repo, name, &suggestions, Some(&source_list));
-    bail!("{error_msg}")
-}
-
-/// Get fuzzy suggestions for overlay names from multi-source config.
-fn get_fuzzy_suggestions_multi_source(
-    manager: &sources::SourceManager,
-    org: &str,
-    repo: &str,
-    query: &str,
-) -> Vec<String> {
-    let available: Vec<String> = manager
-        .list_overlays_for_repo(org, repo)
-        .into_iter()
-        .map(|n| n.to_string())
-        .collect();
-    fuzzy_suggest(query, &available)
-}
-
-/// Find fuzzy matches for a query in the given candidates.
-fn fuzzy_suggest(query: &str, candidates: &[String]) -> Vec<String> {
-    if candidates.is_empty() {
-        return Vec::new();
-    }
-    let matcher = OverlayMatcher::new();
-    matcher.suggest(query, candidates, 3)
-}
-
-/// Format a "not found" error message with optional fuzzy suggestions and source list.
-fn format_not_found_error(
-    org: &str,
-    repo: &str,
-    name: &str,
-    suggestions: &[String],
-    source_list: Option<&str>,
-) -> String {
-    use std::fmt::Write;
-
-    let mut msg = format!("Overlay not found: {org}/{repo}/{name}");
-
-    if let Some(sources) = source_list {
-        let _ = write!(msg, "\n\nSearched sources: {sources}");
-    }
-
-    if !suggestions.is_empty() {
-        msg.push_str("\n\nDid you mean?");
-        for suggestion in suggestions {
-            let _ = write!(msg, "\n  - {suggestion}");
-        }
-    }
-
-    let _ = write!(
-        msg,
-        "\n\nUse `repoverlay list --filter {org}/{repo}` to see available overlays."
-    );
-
-    msg
 }
 
 /// Apply an overlay to a target git repository.
